refactor: split out action plugins to simplify integration
